requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Email handling
email-validator==2.1.0.post1

# LLM - Using Groq (FREE!)
groq==0.11.0

# PDF generation
reportlab==4.0.7

# Utilities
requests==2.31.0
python-dateutil==2.8.2